+++ src/bin/ply_play.rs
@@ -1,21 +1,23 @@
 use cgmath::Point3;
 use clap::Parser;
 use log::{debug, info, trace, warn};
-use std::cell::RefCell;
 use std::ffi::OsString;
-use std::fs::File;
-use std::io::BufReader;
 use std::path::{Path, PathBuf};
 use tempfile::tempdir;
-use vivotk::abr::quetra::{Quetra, QuetraMultiview};
-use vivotk::abr::{RateAdapter, MCKP};
-use vivotk::codec::decoder::{DracoDecoder, NoopDecoder, Tmc2rsDecoder};
-use vivotk::codec::Decoder;
-use vivotk::dash::buffer::{Buffer, FrameStatus};
-use vivotk::dash::fetcher::{FetchResult, Fetcher};
-use vivotk::dash::{ThroughputPrediction, ViewportPrediction};
-use vivotk::formats::pointxyzrgba::PointXyzRgba;
-use vivotk::formats::PointCloud;
+use vivotk::abr::{
+    quetra::{Quetra, QuetraMultiview},
+    RateAdapter, MCKP,
+};
+use vivotk::codec::{
+    decoder::{DracoDecoder, NoopDecoder, Tmc2rsDecoder},
+    Decoder,
+};
+use vivotk::dash::{
+    buffer::{Buffer, FrameStatus},
+    fetcher::{FetchResult, Fetcher},
+    ThroughputPrediction, ViewportPrediction,
+};
+use vivotk::formats::{pointxyzrgba::PointXyzRgba, PointCloud};
 use vivotk::render::wgpu::{
     builder::{EventType, RenderBuilder, RenderEvent},
     camera::{Camera, CameraPosition},
@@ -24,9 +26,10 @@ use vivotk::render::wgpu::{
     reader::{FrameRequest, PcdAsyncReader, RenderReader},
     renderer::Renderer,
 };
+use vivotk::simulation::{CameraTrace, NetworkTrace};
 use vivotk::utils::{
-    get_cosines, predict_quality, read_file_to_point_cloud, ExponentialMovingAverage, LastValue,
-    SimpleRunningAverage, GAEMA, LPEMA,
+    get_cosines, predict_quality, ExponentialMovingAverage, LastValue, SimpleRunningAverage, GAEMA,
+    LPEMA,
 };
 use vivotk::{BufMsg, PCMetadata};
 
@@ -403,140 +406,6 @@ impl From<FetchRequest> for FrameRequest {
             object_id: val.object_id,
             frame_offset: val.frame_offset,
             camera_pos: val.camera_pos,
-        }
-    }
-}
-
-struct NetworkTrace {
-    data: Vec<f64>,
-    index: RefCell<usize>,
-}
-
-impl NetworkTrace {
-    /// The network trace file to contain the network bandwidth in Kbps, each line representing 1 bandwidth sample.
-    /// # Arguments
-    ///
-    /// * `path` - The path to the network trace file.
-    fn new(path: &Path) -> Self {
-        use std::io::BufRead;
-
-        let file = File::open(path).unwrap();
-        let reader = BufReader::new(file);
-        let data = reader
-            .lines()
-            .map(|line| line.unwrap().trim().parse::<f64>().unwrap())
-            .collect();
-        NetworkTrace {
-            data,
-            index: RefCell::new(0),
-        }
-    }
-
-    // Get the next bandwidth sample
-    fn next(&self) -> f64 {
-        let idx = *self.index.borrow();
-        let next_idx = (idx + 1) % self.data.len();
-        *self.index.borrow_mut() = next_idx;
-        self.data[idx]
-    }
-}
-
-struct CameraTrace {
-    data: Vec<CameraPosition>,
-    index: RefCell<usize>,
-    path: PathBuf,
-}
-
-impl CameraTrace {
-    /// The network trace file to contain the network bandwidth in Kbps, each line representing 1 bandwidth sample.
-    /// # Arguments
-    ///
-    /// * `path` - The path to the network trace file.
-    fn new(path: &Path, is_record: bool) -> Self {
-        use std::io::BufRead;
-        match File::open(path) {
-            Err(err) => {
-                if !is_record {
-                    panic!("Failed to open camera trace file: {err:?}");
-                }
-                Self {
-                    data: Vec::new(),
-                    index: RefCell::new(0),
-                    path: path.to_path_buf(),
-                }
-            }
-            Ok(file) => {
-                if is_record {
-                    panic!("Camera trace file already exists: {path:?}");
-                }
-                let reader = BufReader::new(file);
-                let data = reader
-                    .lines()
-                    .map(|line| {
-                        let line = line.unwrap();
-                        let mut it = line.trim().split(',').map(|s| s.parse::<f32>().unwrap());
-                        let position =
-                            Point3::new(it.next().unwrap(), it.next().unwrap(), it.next().unwrap());
-                        let pitch = cgmath::Deg(it.next().unwrap()).into();
-                        let yaw = cgmath::Deg(it.next().unwrap()).into();
-                        CameraPosition {
-                            position,
-                            pitch,
-                            yaw,
-                        }
-                    })
-                    .collect();
-                Self {
-                    data,
-                    index: RefCell::new(0),
-                    path: path.to_path_buf(),
-                }
-            }
-        }
-    }
-
-    /// Get the next bandwidth sample. Used when playing back a camera trace.
-    fn next(&self) -> CameraPosition {
-        let idx = *self.index.borrow();
-        let next_idx = (idx + 1) % self.data.len();
-        *self.index.borrow_mut() = next_idx;
-        self.data[idx]
-    }
-
-    /// Add a new position to the trace. Used when recording a camera trace.
-    fn add(&mut self, pos: CameraPosition) {
-        self.data.push(pos);
-    }
-}
-
-impl Drop for CameraTrace {
-    fn drop(&mut self) {
-        use std::io::BufWriter;
-        use std::io::Write;
-
-        match std::fs::OpenOptions::new()
-            .write(true)
-            .create_new(true)
-            .open(&self.path)
-        {
-            Ok(mut file) => {
-                let mut writer = BufWriter::new(&mut file);
-                for pos in &self.data {
-                    writeln!(
-                        writer,
-                        "{},{},{},{},{},0.0",
-                        pos.position.x,
-                        pos.position.y,
-                        pos.position.z,
-                        pos.pitch.0.to_degrees(),
-                        pos.yaw.0.to_degrees()
-                    )
-                    .unwrap();
-                }
-            }
-            Err(_) => {
-                warn!("Camera trace file already exists, not writing");
-            }
         }
     }
 }
